respond to all pr feedback
