+++ vscode-dotnet-runtime-library/src/Acquisition/VersionResolver.ts
@@ -5,6 +5,7 @@
 import * as semver from 'semver';
 import { IEventStream } from '../EventStream/EventStream';
 import {
+    DotnetOfflineFailure,
     DotnetVersionResolutionCompleted,
     DotnetVersionResolutionError,
 } from '../EventStream/EventStreamEvents';
@@ -12,6 +13,12 @@ import { IExtensionState } from '../IExtensionState';
 import { WebRequestWorker } from '../Utils/WebRequestWorker';
 import { IVersionResolver } from './IVersionResolver';
 import { ReleasesResult } from './ReleasesResult';
+import { DotnetVersionSupportPhase,
+    DotnetVersionSupportStatus,
+    IDotnetListVersionsContext,
+    IDotnetListVersionsResult,
+    IDotnetVersion
+} from '../IDotnetListVersionsContext';
 
 export class VersionResolver implements IVersionResolver {
     protected webWorker: WebRequestWorker;
@@ -21,6 +28,59 @@ export class VersionResolver implements IVersionResolver {
     constructor(extensionState: IExtensionState,
                 private readonly eventStream: IEventStream) {
         this.webWorker = new WebRequestWorker(extensionState, eventStream, this.releasesUrl, this.releasesKey);
+    }
+
+    /**
+     * @remarks
+     * Use the release.json manifest that contains the newest version of the SDK and Runtime for each major.minor of .NET to get the available versions.
+     * Relies on the context listRuntimes to tell if it should get runtime or sdk versions.
+     *
+     * @params
+     * webWorker - This class can use its own web-worker or a custom one for testing purposes.
+     *
+     * @returns
+     * IDotnetListVersionsResult of versions available.
+     *
+     * @throws
+     * Exception if the API service for releases-index.json is unavailable.
+    */
+    public async GetAvailableDotnetVersions(commandContext: IDotnetListVersionsContext | undefined, webWorker: WebRequestWorker | undefined) : Promise<IDotnetListVersionsResult>
+    {
+        // If shouldObtainSdkVersions === false, get Runtimes. Else, get Sdks.
+        const shouldObtainSdkVersions : boolean = !commandContext?.listRuntimes;
+        const availableVersions : IDotnetListVersionsResult = [];
+
+        const response = webWorker != null ? await webWorker.getCachedData() : await this.webWorker.getCachedData();
+
+        return new Promise<IDotnetListVersionsResult>((resolve, reject) =>
+        {
+            if (!response)
+            {
+                const offlineError = new Error('Unable to connect to the index server: Cannot find .NET versions.');
+                this.eventStream.post(new DotnetOfflineFailure(offlineError, "any"));
+                reject(offlineError);
+            }
+            else
+            {
+                const sdkDetailsJson = JSON.parse(response)['releases-index'];
+
+                for(const availableSdk of sdkDetailsJson)
+                {
+                    if(availableSdk['release-type'] === 'lts' || availableSdk['release-type'] === 'sts')
+                    {
+                        availableVersions.push({
+                                supportStatus: (availableSdk['release-type'] as DotnetVersionSupportStatus),
+                                supportPhase: (availableSdk['support-phase'] as DotnetVersionSupportPhase),
+                                version: availableSdk[shouldObtainSdkVersions ? 'latest-sdk' : 'latest-runtime'],
+                                channelVersion: availableSdk['channel-version']
+                            } as IDotnetVersion
+                        );
+                    }
+                }
+            }
+
+            resolve(availableVersions);
+        });
     }
 
     public async getFullRuntimeVersion(version: string): Promise<string> {
